Version 1.1
Now the script will also cope with coma separated list that is not enclosed in brackets, which is common in Python code.

for |a|, b in some_dict.items()  -->    for b, |a| in some_dict.items()

Moreover, I have made the code (with full suite of tests) available on code.google.com. Link is on my blog: http://blog.kamil.dworakowski.name/2008/01/smarter-swap-parameters-for-vim.html.